Use PImpl for TagThrottler class
